next-authを使った自動セッションの取得
