+++ app/admin/login/page.tsx
@@ -1,42 +1,18 @@
-"use client";
+import { auth } from "@/auth";
+import { redirect } from "next/navigation";
+import LoginForm from "./LoginForm";
 
-import { useState } from "react";
-import { signIn } from "next-auth/react";
-import { useRouter } from "next/navigation";
+export default async function AuthLogin() {
+  // cookieからセッションを自動取得して検証
+  const session = await auth();
 
-export default function AuthLogin() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
-  const router = useRouter();
-
-  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    setError("");
-    setIsLoading(true);
-
-    try {
-      const result = await signIn("credentials", {
-        email,
-        password,
-        redirect: false,
-      });
-
-      if (result?.error) {
-        setError("メールアドレスまたはパスワードが正しくありません");
-        setIsLoading(false);
-      } else if (result?.ok) {
-        // ログイン成功
-        router.push("/admin"); // 管理者ページにリダイレクト
-        router.refresh();
-      }
-    } catch (error) {
-      console.error("ログインエラー:", error);
-      setError("ログインに失敗しました");
-      setIsLoading(false);
+  // 既にログインしている場合、ダッシュボードにリダイレクト
+  if (session) {
+    const userRole = (session.user as { role?: string })?.role;
+    if (userRole === "admin") {
+      redirect("/admin/dashboard");
     }
-  };
+  }
 
   return (
     <div className="flex min-h-screen items-center justify-center bg-gray-50">
@@ -46,62 +22,7 @@ export default function AuthLogin() {
             管理者ログイン
           </h2>
         </div>
-        <form className="mt-8 space-y-6" onSubmit={handleSubmit}>
-          {error && (
-            <div className="rounded-md bg-red-50 p-4">
-              <p className="text-sm text-red-800">{error}</p>
-            </div>
-          )}
-          <div className="space-y-4">
-            <div>
-              <label
-                htmlFor="email"
-                className="block text-sm font-medium text-gray-700"
-              >
-                メールアドレス
-              </label>
-              <input
-                id="email"
-                name="email"
-                type="email"
-                required
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                className="mt-1 block w-full rounded-md border border-gray-300 px-3 py-2 text-gray-900 placeholder-gray-400 focus:border-[#b43353] focus:outline-none focus:ring-1 focus:ring-[#b43353]"
-                placeholder="admin@example.com"
-                disabled={isLoading}
-              />
-            </div>
-            <div>
-              <label
-                htmlFor="password"
-                className="block text-sm font-medium text-gray-700"
-              >
-                パスワード
-              </label>
-              <input
-                id="password"
-                name="password"
-                type="password"
-                required
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                className="mt-1 block w-full rounded-md border border-gray-300 px-3 py-2 text-gray-900 placeholder-gray-400 focus:border-[#b43353] focus:outline-none focus:ring-1 focus:ring-[#b43353]"
-                placeholder="パスワードを入力"
-                disabled={isLoading}
-              />
-            </div>
-          </div>
-          <div>
-            <button
-              type="submit"
-              disabled={isLoading}
-              className="group relative flex w-full justify-center rounded-md bg-[#b43353] px-4 py-2 text-sm font-semibold text-white hover:bg-[#9a2a45] focus:outline-none focus:ring-2 focus:ring-[#b43353] focus:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50"
-            >
-              {isLoading ? "ログイン中..." : "ログイン"}
-            </button>
-          </div>
-        </form>
+        <LoginForm />
       </div>
     </div>
   );
